+++ src/main/deploy/Subsystems.json
@@ -1,6 +1,6 @@
 {
     "swerve": true,
     "lights": false,
-    "elevator": true,
+    "elevator": false,
     "vision": true
 }
