+++ src/main/java/frc/robot/Konstants.java
@@ -1,44 +1,361 @@
 package frc.robot;
 
 import static edu.wpi.first.units.Units.Amps;
+import static edu.wpi.first.units.Units.Inches;
+import static edu.wpi.first.units.Units.KilogramSquareMeters;
+import static edu.wpi.first.units.Units.MetersPerSecond;
+import static edu.wpi.first.units.Units.RadiansPerSecond;
+import static edu.wpi.first.units.Units.Rotations;
+import static edu.wpi.first.units.Units.RotationsPerSecond;
+import static edu.wpi.first.units.Units.Volts;
+import static frc.robot.Konstants.SwerveConstants.kPigeonID;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 
-import com.ctre.phoenix6.CANBus;
+import com.ctre.phoenix6.CANBus;    
+import com.ctre.phoenix6.configs.CANcoderConfiguration;
+import com.ctre.phoenix6.configs.CurrentLimitsConfigs;
+import com.ctre.phoenix6.configs.Pigeon2Configuration;
 import com.ctre.phoenix6.configs.Slot0Configs;
+import com.ctre.phoenix6.configs.TalonFXConfiguration;
+import com.ctre.phoenix6.hardware.CANcoder;
+import com.ctre.phoenix6.hardware.TalonFX;
+import com.ctre.phoenix6.signals.StaticFeedforwardSignValue;
+import com.ctre.phoenix6.swerve.SwerveDrivetrain;
+import com.ctre.phoenix6.swerve.SwerveDrivetrainConstants;
+import com.ctre.phoenix6.swerve.SwerveModuleConstants;
 import com.ctre.phoenix6.swerve.SwerveModuleConstants.ClosedLoopOutputType;
+import com.ctre.phoenix6.swerve.SwerveModuleConstants.DriveMotorArrangement;
+import com.ctre.phoenix6.swerve.SwerveModuleConstants.SteerFeedbackType;
+import com.ctre.phoenix6.swerve.SwerveModuleConstants.SteerMotorArrangement;
+import com.ctre.phoenix6.swerve.SwerveModuleConstantsFactory;
 import com.pathplanner.lib.config.PIDConstants;
 import com.pathplanner.lib.controllers.PPHolonomicDriveController;
 
-import edu.wpi.first.math.geometry.Pose2d;
-import edu.wpi.first.math.geometry.Rotation2d;
-import edu.wpi.first.math.util.Units;
+import edu.wpi.first.math.Matrix;
+import edu.wpi.first.math.numbers.N1;
+import edu.wpi.first.math.numbers.N3;
+import edu.wpi.first.units.measure.Angle;
 import edu.wpi.first.units.measure.Current;
+import edu.wpi.first.units.measure.Distance;
 import edu.wpi.first.units.measure.LinearVelocity;
 //import frc.robot.subsystems.swerve.SwerveConstantsConfigurator;
+import frc.robot.preferences.SKPreferences;
 
-import static edu.wpi.first.units.Units.MetersPerSecond;
-import static edu.wpi.first.units.Units.Rotations;
-import static edu.wpi.first.units.Units.Radians;
+//import static edu.wpi.first.units.Units.Rotations;
+//import static edu.wpi.first.units.Units.Radians;
+//import edu.wpi.first.units.measure.Distance;
+//import frc.robot.subsystems.swerve.SwerveConstantsConfigurator;
+//import edu.wpi.first.math.util.Units;
+import com.pathplanner.lib.config.PIDConstants;
+import com.revrobotics.spark.SparkClosedLoopController;
+import com.revrobotics.spark.config.ClosedLoopConfig.FeedbackSensor;
+import com.revrobotics.spark.config.LimitSwitchConfig.Type;
+import com.revrobotics.spark.config.SparkBaseConfig.IdleMode;
+import com.revrobotics.spark.config.SparkFlexConfig;
+import edu.wpi.first.units.measure.MomentOfInertia;
+import edu.wpi.first.units.measure.Voltage;
+import frc.robot.subsystems.SKSwerve;
 
+import edu.wpi.first.math.geometry.Rotation2d;
+import edu.wpi.first.math.geometry.Pose2d;
+import edu.wpi.first.math.util.Units;
+import edu.wpi.first.units.measure.Angle;
 
 //import edu.wpi.first.units.measure.Distance;
 
 public final class Konstants
 {
 
+    // Generated by the Tuner X Swerve Project Generator
+// https://v6.docs.ctr-electronics.com/en/stable/docs/tuner/tuner-swerve/index.html
+public class TunerConstants {
+    // Both sets of gains need to be tuned to your individual robot.
+    
+    // The steer motor uses any SwerveModule.SteerRequestType control request with the
+    // output type specified by SwerveModuleConstants.SteerMotorClosedLoopOutput
+    private static final Slot0Configs steerGains = new Slot0Configs()
+    .withKP(100).withKI(0).withKD(0.5)
+    .withKS(0.1).withKV(2.66).withKA(0)
+    .withStaticFeedforwardSign(StaticFeedforwardSignValue.UseClosedLoopSign);
+    // When using closed-loop control, the drive motor uses the control
+    // output type specified by SwerveModuleConstants.DriveMotorClosedLoopOutput
+    private static final Slot0Configs driveGains = new Slot0Configs()
+    .withKP(0.1).withKI(0).withKD(0)
+        .withKS(0).withKV(0.124);
+        
+    // The closed-loop output type to use for the steer motors;
+    // This affects the PID/FF gains for the steer motors
+    private static final ClosedLoopOutputType kSteerClosedLoopOutput = ClosedLoopOutputType.Voltage;
+    // The closed-loop output type to use for the drive motors;
+    // This affects the PID/FF gains for the drive motors
+    private static final ClosedLoopOutputType kDriveClosedLoopOutput = ClosedLoopOutputType.Voltage;
+    
+    // The type of motor used for the drive motor
+    private static final DriveMotorArrangement kDriveMotorType = DriveMotorArrangement.TalonFX_Integrated;
+    // The type of motor used for the drive motor
+    private static final SteerMotorArrangement kSteerMotorType = SteerMotorArrangement.TalonFX_Integrated;
+
+    // The remote sensor feedback type to use for the steer motors;
+    // When not Pro-licensed, Fused*/Sync* automatically fall back to Remote*
+    private static final SteerFeedbackType kSteerFeedbackType = SteerFeedbackType.FusedCANcoder;
+    
+    // The stator current at which the wheels start to slip;
+    // This needs to be tuned to your individual robot
+    private static final Current kSlipCurrent = Amps.of(120.0);
+    
+    // Initial configs for the drive and steer motors and the azimuth encoder; these cannot be null.
+    // Some configs will be overwritten; check the `with*InitialConfigs()` API documentation.
+    private static final TalonFXConfiguration driveInitialConfigs = new TalonFXConfiguration()
+    .withCurrentLimits(
+        new CurrentLimitsConfigs()
+        .withSupplyCurrentLimit(70.0) // Amps
+        .withSupplyCurrentLowerLimit(50) // Amps
+        .withSupplyCurrentLowerTime(0.5) // Seconds
+            .withSupplyCurrentLimitEnable(true)
+            
+            .withStatorCurrentLimit(60) // Amps
+            .withStatorCurrentLimitEnable(true)
+            );
+
+    private static final TalonFXConfiguration steerInitialConfigs = new TalonFXConfiguration()
+    .withCurrentLimits(
+        new CurrentLimitsConfigs()
+        // Swerve azimuth does not require much torque output, so we can set a relatively low
+                // stator current limit to help avoid brownouts without impacting performance.
+                .withStatorCurrentLimit(Amps.of(60))
+                .withStatorCurrentLimitEnable(true)
+        );
+    private static final CANcoderConfiguration encoderInitialConfigs = new CANcoderConfiguration();
+    // Configs for the Pigeon 2; leave this null to skip applying Pigeon 2 configs
+    private static final Pigeon2Configuration pigeonConfigs = null;
+
+    // CAN bus that the devices are located on;
+    // All swerve devices must share the same CAN bus
+    public static final CANBus kCANBus = new CANBus("SwerveCANivore", "./logs/example.hoot");
+    
+    // Theoretical free speed (m/s) at 12 V applied output;
+    // This needs to be tuned to your individual robot
+    public static final LinearVelocity kSpeedAt12Volts = MetersPerSecond.of(4.73);
+    
+    public static double MaxSpeed = kSpeedAt12Volts.in(MetersPerSecond); // kSpeedAt12Volts desired top speed
+    public static double MaxAngularRate = RotationsPerSecond.of(0.75).in(RadiansPerSecond); // 3/4 of a rotation per second max angular velocity
+    
+    // Every 1 rotation of the azimuth results in kCoupleRatio drive motor turns;
+    // This may need to be tuned to your individual robot
+    private static final double kCoupleRatio = 3.5714285714285716;
+    
+    private static final double kDriveGearRatio = 6.746031746031747;
+    private static final double kSteerGearRatio = 21.428571428571427;
+    private static final Distance kWheelRadius = Inches.of(2);
+
+    private static final boolean kInvertLeftSide = true;
+    private static final boolean kInvertRightSide = false;
+
+    private static final int kPigeonId = kPigeonID;
+    
+    // These are only used for simulation
+    private static final MomentOfInertia kSteerInertia = KilogramSquareMeters.of(0.01);
+    private static final MomentOfInertia kDriveInertia = KilogramSquareMeters.of(0.01);
+    // Simulated voltage necessary to overcome friction
+    private static final Voltage kSteerFrictionVoltage = Volts.of(0.2);
+    private static final Voltage kDriveFrictionVoltage = Volts.of(0.2);
+    
+    public static final SwerveDrivetrainConstants DrivetrainConstants = new SwerveDrivetrainConstants()
+    .withCANBusName(kCANBus.getName())
+    .withPigeon2Id(kPigeonId)
+    .withPigeon2Configs(pigeonConfigs);
+    
+    private static final SwerveModuleConstantsFactory<TalonFXConfiguration, TalonFXConfiguration, CANcoderConfiguration> ConstantCreator =
+    new SwerveModuleConstantsFactory<TalonFXConfiguration, TalonFXConfiguration, CANcoderConfiguration>()
+    .withDriveMotorGearRatio(kDriveGearRatio)
+    .withSteerMotorGearRatio(kSteerGearRatio)
+    .withCouplingGearRatio(kCoupleRatio)
+    .withWheelRadius(kWheelRadius)
+    .withSteerMotorGains(steerGains)
+            .withDriveMotorGains(driveGains)
+            .withSteerMotorClosedLoopOutput(kSteerClosedLoopOutput)
+            .withDriveMotorClosedLoopOutput(kDriveClosedLoopOutput)
+            .withSlipCurrent(kSlipCurrent)
+            .withSpeedAt12Volts(kSpeedAt12Volts)
+            .withDriveMotorType(kDriveMotorType)
+            .withSteerMotorType(kSteerMotorType)
+            .withFeedbackSource(kSteerFeedbackType)
+            .withDriveMotorInitialConfigs(driveInitialConfigs)
+            .withSteerMotorInitialConfigs(steerInitialConfigs)
+            .withEncoderInitialConfigs(encoderInitialConfigs)
+            .withSteerInertia(kSteerInertia)
+            .withDriveInertia(kDriveInertia)
+            .withSteerFrictionVoltage(kSteerFrictionVoltage)
+            .withDriveFrictionVoltage(kDriveFrictionVoltage);
+            
+            
+            // Front Left
+            private static final int kFrontLeftDriveMotorId = 1;
+            private static final int kFrontLeftSteerMotorId = 11;
+            private static final int kFrontLeftEncoderId = 21;
+            private static final Angle kFrontLeftEncoderOffset = Rotations.of(-0.184326171875);
+            private static final boolean kFrontLeftSteerMotorInverted = true;
+    private static final boolean kFrontLeftEncoderInverted = false;
+    
+    private static final Distance kFrontLeftXPos = Inches.of(10.8125);
+    private static final Distance kFrontLeftYPos = Inches.of(10.8125);
+
+    // Front Right
+    private static final int kFrontRightDriveMotorId = 2;
+    private static final int kFrontRightSteerMotorId = 12;
+    private static final int kFrontRightEncoderId = 22;
+    private static final Angle kFrontRightEncoderOffset = Rotations.of(0.113525390625);
+    private static final boolean kFrontRightSteerMotorInverted = true;
+    private static final boolean kFrontRightEncoderInverted = false;
+    
+    private static final Distance kFrontRightXPos = Inches.of(10.8125);
+    private static final Distance kFrontRightYPos = Inches.of(-10.8125);
+    
+    // Back Left
+    private static final int kBackLeftDriveMotorId = 3;
+    private static final int kBackLeftSteerMotorId = 13;
+    private static final int kBackLeftEncoderId = 23;
+    private static final Angle kBackLeftEncoderOffset = Rotations.of(-0.036865234375);
+    private static final boolean kBackLeftSteerMotorInverted = true;
+    private static final boolean kBackLeftEncoderInverted = false;
+    
+    private static final Distance kBackLeftXPos = Inches.of(-10.8125);
+    private static final Distance kBackLeftYPos = Inches.of(10.8125);
+    
+    // Back Right
+    private static final int kBackRightDriveMotorId = 4;
+    private static final int kBackRightSteerMotorId = 14;
+    private static final int kBackRightEncoderId = 24;
+    private static final Angle kBackRightEncoderOffset = Rotations.of(0.441162109375);
+    private static final boolean kBackRightSteerMotorInverted = true;
+    private static final boolean kBackRightEncoderInverted = false;
+    
+    private static final Distance kBackRightXPos = Inches.of(-10.8125);
+    private static final Distance kBackRightYPos = Inches.of(-10.8125);
+    
+    
+    public static final SwerveModuleConstants<TalonFXConfiguration, TalonFXConfiguration, CANcoderConfiguration> FrontLeft =
+    ConstantCreator.createModuleConstants(
+        kFrontLeftSteerMotorId, kFrontLeftDriveMotorId, kFrontLeftEncoderId, kFrontLeftEncoderOffset,
+        kFrontLeftXPos, kFrontLeftYPos, kInvertLeftSide, kFrontLeftSteerMotorInverted, kFrontLeftEncoderInverted
+        );
+        public static final SwerveModuleConstants<TalonFXConfiguration, TalonFXConfiguration, CANcoderConfiguration> FrontRight =
+        ConstantCreator.createModuleConstants(
+            kFrontRightSteerMotorId, kFrontRightDriveMotorId, kFrontRightEncoderId, kFrontRightEncoderOffset,
+            kFrontRightXPos, kFrontRightYPos, kInvertRightSide, kFrontRightSteerMotorInverted, kFrontRightEncoderInverted
+            );
+            public static final SwerveModuleConstants<TalonFXConfiguration, TalonFXConfiguration, CANcoderConfiguration> BackLeft =
+            ConstantCreator.createModuleConstants(
+                kBackLeftSteerMotorId, kBackLeftDriveMotorId, kBackLeftEncoderId, kBackLeftEncoderOffset,
+                kBackLeftXPos, kBackLeftYPos, kInvertLeftSide, kBackLeftSteerMotorInverted, kBackLeftEncoderInverted
+                );
+                public static final SwerveModuleConstants<TalonFXConfiguration, TalonFXConfiguration, CANcoderConfiguration> BackRight =
+        ConstantCreator.createModuleConstants(
+            kBackRightSteerMotorId, kBackRightDriveMotorId, kBackRightEncoderId, kBackRightEncoderOffset,
+            kBackRightXPos, kBackRightYPos, kInvertRightSide, kBackRightSteerMotorInverted, kBackRightEncoderInverted
+            );
+
+    /**
+     * Creates a CommandSwerveDrivetrain instance.
+     * This should only be called once in your robot program,.
+     */
+    public static SKSwerve createDrivetrain() {
+        return new SKSwerve(
+            DrivetrainConstants, FrontLeft, FrontRight, BackLeft, BackRight
+            );
+    }
+
+
+    /**
+     * Swerve Drive class utilizing CTR Electronics' Phoenix 6 API with the selected device types.
+     */
+    public static class TunerSwerveDrivetrain extends SwerveDrivetrain<TalonFX, TalonFX, CANcoder> {
+        /**
+         * Constructs a CTRE SwerveDrivetrain using the specified constants.
+         * <p>
+         * This constructs the underlying hardware devices, so users should not construct
+         * the devices themselves. If they need the devices, they can access them through
+         * getters in the classes.
+         *
+         * @param drivetrainConstants   Drivetrain-wide constants for the swerve drive
+         * @param modules               Constants for each specific module
+         */
+        public TunerSwerveDrivetrain(
+            SwerveDrivetrainConstants drivetrainConstants,
+            SwerveModuleConstants<?, ?, ?>... modules
+        ) {
+            super(
+                TalonFX::new, TalonFX::new, CANcoder::new,
+                drivetrainConstants, modules
+            );
+        }
+
+        /**
+         * Constructs a CTRE SwerveDrivetrain using the specified constants.
+         * <p>
+         * This constructs the underlying hardware devices, so users should not construct
+         * the devices themselves. If they need the devices, they can access them through
+         * getters in the classes.
+         *
+         * @param drivetrainConstants     Drivetrain-wide constants for the swerve drive
+         * @param odometryUpdateFrequency The frequency to run the odometry loop. If
+         *                                unspecified or set to 0 Hz, this is 250 Hz on
+         *                                CAN FD, and 100 Hz on CAN 2.0.
+         * @param modules                 Constants for each specific module
+         */
+        public TunerSwerveDrivetrain(
+            SwerveDrivetrainConstants drivetrainConstants,
+            double odometryUpdateFrequency,
+            SwerveModuleConstants<?, ?, ?>... modules
+        ) {
+            super(
+                TalonFX::new, TalonFX::new, CANcoder::new,
+                drivetrainConstants, odometryUpdateFrequency, modules
+            );
+        }
+
+        /**
+         * Constructs a CTRE SwerveDrivetrain using the specified constants.
+         * <p>
+         * This constructs the underlying hardware devices, so users should not construct
+         * the devices themselves. If they need the devices, they can access them through
+         * getters in the classes.
+         *
+         * @param drivetrainConstants       Drivetrain-wide constants for the swerve drive
+         * @param odometryUpdateFrequency   The frequency to run the odometry loop. If
+         *                                  unspecified or set to 0 Hz, this is 250 Hz on
+         *                                  CAN FD, and 100 Hz on CAN 2.0.
+         * @param odometryStandardDeviation The standard deviation for odometry calculation
+         *                                  in the form [x, y, theta]ᵀ, with units in meters
+         *                                  and radians
+         * @param visionStandardDeviation   The standard deviation for vision calculation
+         *                                  in the form [x, y, theta]ᵀ, with units in meters
+         *                                  and radians
+         * @param modules                   Constants for each specific module
+         */
+        public TunerSwerveDrivetrain(
+            SwerveDrivetrainConstants drivetrainConstants,
+            double odometryUpdateFrequency,
+            Matrix<N3, N1> odometryStandardDeviation,
+            Matrix<N3, N1> visionStandardDeviation,
+            SwerveModuleConstants<?, ?, ?>... modules
+        ) {
+            super(
+                TalonFX::new, TalonFX::new, CANcoder::new,
+                drivetrainConstants, odometryUpdateFrequency,
+                odometryStandardDeviation, visionStandardDeviation, modules
+            );
+        }
+    }
+}
+
     public static final class SwerveConstants
     {
-        //TODO: Merge TunerConstants with this?
-
-
         //Device Settings and Default States
-
-
-
         //pigeon ID
         public static final int kPigeonID = 30; //30
 
@@ -59,206 +376,14 @@ public final class Konstants
         public static final int kBackLeftEncoderID = 23; //23
         public static final int kBackRightEncoderID = 24; //24
 
-        //The offset of the encoders in radians
-        //fl 0.35
-        //fr 2.225
-        //bl 1.35
-        //br 4.353
-
-        public static final Double kFrontLeftEncoderOffsetRadians = (0.35); //old offset, don't delete: -0.184326171875;
-        public static final Double kFrontRightEncoderOffsetRadians = (2.225); //old offset, don't delete: 0.113525390625;
-        public static final Double kBackLeftEncoderOffsetRadians = (1.35); //old offset, don't delete: -0.036865234375;
-        public static final Double kBackRightEncoderOffsetRadians = (4.353); //old offset, don't delete: 0.441162109375;
-
-        //determines if the encoders are inverted
-        public static final boolean kIsFrontLeftEncoderInverted = false;
-        public static final boolean kIsFrontRightEncoderInverted = false;
-        public static final boolean kIsBackLeftEncoderInverted = false;
-        public static final boolean kIsBackRightEncoderInverted = false;
-
-        //constants which determine if the drive motors are inverted, negative if they are inverted, positive if not.
-        public static final double kFrontLeftDriveInverted = -1.0;
-        public static final double kFrontRightDriveInverted = -1.0;
-        public static final double kBackLeftDriveInverted = -1.0;
-        public static final double kBackRightDriveInverted = -1.0;
-
-        //if the sides are inverted
-        public static final boolean kInvertLeftSide = false;
-        public static final boolean kInvertRightSide = true;
-
-        //if the turn motors are inverted
-        public static final boolean kTurnMotorsReversed = true;
-
-
-
-        //Robot Dimension values
-
-
-
-        //swerve chassis width and length in inches 
+        /**Swerve chassis width and length in inches*/ 
         public static final int kChassisLength = 27;
-        public static final int kChassisWidth = 27;
 
-        // "Front-to-back Encoder Distance in inches"
-        public static final double kFrontToBackEncoderDistInches = 21.625;
-        //"Left-to-right Encoder Distance in inches"
-        public static final double kLeftToRightEncoderDistInches = 21.625;
+        /**The deadzone on the controller's joysticks*/
+        public static final double kJoystickDeadband = 0.2;
 
-        // Module positions for kinematics, distances between encoders divided by two
-        // Front left                                                                      //TODO: decide on module position in x and y or translation 2ds
-        public static final double kFrontLeftXPos = kLeftToRightEncoderDistInches / 2;
-        public static final double kFrontLeftYPos = kFrontToBackEncoderDistInches / 2;
-        // Front right
-        public static final double kFrontRightXPos = kLeftToRightEncoderDistInches / 2;
-        public static final double kFrontRightYPos = -kFrontToBackEncoderDistInches / 2;
-        // Back left
-        public static final double kBackLeftXPos = -kLeftToRightEncoderDistInches / 2;
-        public static final double kBackLeftYPos = kFrontToBackEncoderDistInches / 2;
-        // Back right
-        public static final double kBackRightXPos = -kLeftToRightEncoderDistInches / 2;
-        public static final double kBackRightYPos = -kFrontToBackEncoderDistInches / 2;
-
-        //radius of the wheels in inches
-        public static final Double kWheelRadiusInches = 2.0;   //inches
-        // Combines a margin of error with the known wheel radius 
-        // to determine an optimized value for odometry
-        public static final double kWheelErrorMargin = 0.0; //inches
-        public static final double kWheelRadius = (kWheelRadiusInches + kWheelErrorMargin); //inches
-
-        //radius of the wheels in meters. One meter is equal to 39.37 inches.
-        private static final Double kWheelRadiusMeters = kWheelRadius  / 39.37;   //meters
-        /**Circumference of the swerve wheels for the drive conversion 
-         * (circumfrance of the wheel times rotations yeilds distance travelled) */
-        public static final Double kWheelCircumferenceMeters = 2 * Math.PI * kWheelRadiusMeters;   //meters
-
-
-        // Every 1 rotation of the azimuth results in kCoupleRatio drive motor turns;
-        // This may need to be tuned to your individual robot
-        //Old
-        public static final double kCoupleGearRatio = 3.5714285714285716;
-
-        //The gear ratios of the drive and turn motors
-        //Old
-        public static final double kDriveGearRatio = 6.746031746031747;
-        public static final double kTurnGearRatio = 21.428571428571427;
-
-
-
-        //PID values
-
-
-
-        //PID Constants for wheels from manual tunning
-        public static final double kDriveP = 0.5;
-        public static final double kDriveI = 0.0;
-        public static final double kDriveD = 0.0;
-        public static final double kDriveS = 0.25;
-        public static final double kDriveV = 0.12;
-        public static final double kDriveA = 0.01;
-
-
-        //rotation controller PID values
-        public static final double kPRotationController = 8.0;
-        public static final double kIRotationController = 0.0;
-        public static final double kDRotationController = 0.2;
-
-        //hold controller PID values
-        public static final double kPHoldController = 12.0;
-        public static final double kIHoldController = 0.0;
-        public static final double kDHoldController = 0.0;
-
-         // The closed-loop output type to use for the steer motors;
-        // This affects the PID/FF gains for the steer motors
-        //Closed loop control uses feeback to control outputs of a dynamic system.
-        public static final ClosedLoopOutputType kTurnClosedLoopOutput = ClosedLoopOutputType.Voltage;
-
-        // The closed-loop output type to use for the drive motors;
-        // This affects the PID/FF gains for the drive motors
-        //Closed loop control uses feeback to control outputs of a dynamic system.
-        public static final ClosedLoopOutputType kDriveClosedLoopOutput = ClosedLoopOutputType.Voltage;
-
-        /**
-         * The error tolerance for the PID controllers of the wheels in radians
-         */
-        public static final double kPIDControllerToleranceDegrees = 6.0;   //TODO: change this to radians?
-
-        public static final double kRotationToleranceRadians = (Math.PI / 360); // rads
-
-
-
-        //Robot Speed and Position Measurements
-        
-
-
-        /**The velocity limit for the swerve drive modules.*/
-        public static final double kMaxVelocityMetersPerSecond = 1.0;  // m/s
-
-        // Theoretical free speed (m/s) at 12v applied output;
-        // This needs to be tuned to your individual robot
-        //Old
-        public static final LinearVelocity kSpeedAt12VoltsMeterPerSecond = MetersPerSecond.of(4.73);  // m/s
-
-        /**The maximum alowed angular speed of the swerve module's motors in degrees per second.*/
-        public static final double kMaxModuleAngularSpeedDegreesPerSecond = 360;
-
-        /** The max speed (m/s) the drive wheels should be allowed to go */
-         public static final double kMaxDriveSpeedMetersPerSecond = 3.0;  // m/s
-        /** The max rotation speed the turn wheels should be allowed to go */
-         public static final double kMaxRotationDegreesPerSecond = 360.0;  // degrees/second
-
-        public static final double kMaxAngularRate = 1.5 * Math.PI;
-        public static final double kMaxAngularVelocity = 2 * Math.PI; // rad/s
-        public static final double kMaxAngularAcceleration = Math.pow(kMaxAngularVelocity, 2); // rad/s^2
-
-        // The inertia expereinced by the robot when attempting to drive or turn.
-        //These are only used for simulation
-        public static final double kSteerInertia = 0.00001;
-        public static final double kDriveInertia = 0.001;
-
-        //Starting positions of the robot when facing toward the opposing alliance's direver station. 
-        public static final Rotation2d kBlueAlliancePerspective = Rotation2d.fromDegrees(0);
-        public static final Rotation2d kRedAlliancePerspective = Rotation2d.fromDegrees(180);
-     
-
-        
-        //Current Limits
-
-
-
-        /**The current limit of the turning motors. This number should be relativley low in comparison 
-        to the drive motor amperage since rotaing dosn't require nearly as much voltage as driving.*/
-        public static final Current kTurningCurrentLimitAmps = Amps.of(60);
-
-        /** Weather the current limits on the turn motors of the swerve drive are enabled. */
-        public static final boolean kTurningCurrentLimitsEnabled = true;
-
-        /**The stator current at which the wheels start to slip.
-        *This needs to be tuned to your individual robot*/
-        public static final double kSlipCurrentAmps = 120;            //TODO: Tune SlipCurrent
-
-        // Simulated voltage necessary to overcome friction
-        public static final double kTurnFrictionVoltage = 0.25;
-        public static final double kDriveFrictionVoltage = 0.25;
-
-
-
-        //Other Constants
-
-
-
-        //Object containing all the neccessary coonstant values for the Swerve subsystem
-        //public static final SwerveConstantsConfigurator config = new SwerveConstantsConfigurator();
-
-        //the deadzone on the controller's joysticks
-        public static final double kJoystickDeadband = 0.2;   //TODO: find approperiate deaband
-        
-        //CANivore name in Phoneix Tuner X for assigning CANbus names
-        public static final String kCANivoreNameString = "SwerveCANivore";
-        public static final CANBus kCANivoreNameCANBus = new CANBus(kCANivoreNameString);
-
-        /** The length of the simulation loop for telemetry/logging. 
-         * Affects the steering and driving inertia.*/
-        public static final double kSimulationLoopPeriod = 0.005;
+        /** The percentage of speed to drive in slow mode. Currently 30% of defualt speed.*/
+        public static final double kSlowModePercentage = 0.3;
     }
 
 
@@ -474,6 +599,8 @@ public final class Konstants
         public static final double kExampleSpeed = 0.5;  //percentage based where 1.0 is max power and 0.0 is minimum
     }
     
+    public static final String kCANivoreName = "SwerveCANivore";
+
     /** The file that is used for system instantiation at runtime */
     public static final String SUBSYSTEMFILE = "Subsystems.json";
 }
